fixzit for .net
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7400 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/WordFilter.h
@@ -248,7 +248,7 @@ inline int WordFilter::filterCharacters(char *input, unsigned int start, size_t 
   }
 
   int randomCharPos, previousCharPos = -1;
-  int maxFilterChar = filterChars.size();
+  int maxFilterChar = (int)filterChars.size();
   int count=0;
   for (unsigned int j=0; j < (unsigned int)length; j++) {
     char c = input[start + j];
